requirements.txt
# Backend dependencies - Python 3.13 compatible
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.26.2
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==0.8.11
transformers==4.35.2
torch==2.1.1
flask==3.0.0
flask-cors==4.0.0

# Authentication and storage
firebase-admin==6.2.0
python-dotenv==1.0.0

# Blockchain integration
aptos-sdk==0.7.1

# Plagiarism detection algorithms
nltk==3.8.1
google-api-python-client==2.108.0
google-auth==2.23.4

# Additional dependencies for compatibility
typing-extensions==4.8.0
setuptools>=68.0.0
wheel>=0.41.0